+++ core/docs/changelog.md
@@ -1,6 +1,6 @@
 # TBD
 ### Changes
-* Interate testsuite' tests in alphabetical order
+* Iterate testsuite's tests in alphabetical order
 
 # 1.19.2
 ### Features